Grouped common files toghether
- grouped constants for colours, dimensions and strings into one
constants file called "assets.js"

- removed each individual folders for components since we have
not used any specific styling sheets or more logic for each component

- moved the `translationObj` to be grouped with its corresponding
helper class, `TranslationTools`
