--- assets/dimensions/dimensions.js
@@ -1,56 +1,0 @@
-export const DefaultDims = {
-    fontSize: 12,
-    paddingSize: 5,
-    pos: 0,
-    length: 0,
-    borderWidth: 3,
-    lineSpacing: 1
-};
-
-export const GraphDims = Object.freeze({
-    upperGraphWidth: 800,
-    upperGraphHeight: 600,
-    upperGraphLeft: 60,
-    upperGraphRight: 400,
-    upperGraphTop: 60,
-    upperGraphBottom: 60,
-    upperGraphTooltipMinWidth: 140,
-    upperGraphTooltipFontSize: 12,
-    upperGraphTooltipTopPadding: 8,
-    upperGraphTooltipLeftPadding: 10,
-    upperGraphTooltipLineSpacing: 3,
-    upperGraphInfoBoxWidth: 240,
-    upperGraphInfoBoxHeight: 200,
-    upperGraphInfoBoxFontSize: 14,
-    upperGraphInfoBoxBorderWidth: 10,
-    upperGraphInfoBoxPaddingLeft: 10,
-    upperGraphInfoBoxLineSpacing: 5,
-    upperGraphInfoBoxLeftMargin: 30,
-    upperGraphChartHeadingFontSize: 20,
-    lowerGraphWidth: 700,
-    lowerGraphHeight: 700,
-    lowerGraphLeft: 60,
-    lowerGraphRight: 400,
-    lowerGraphTop: 60,
-    lowerGraphBottom: 60,
-    lowerGraphArcRadius: 65,
-    lowerGraphCenterArcRadius: 45,
-    lowerGraphCenterArcInnerRadius: 18,
-    lowerGraphCenterArcMargin: 1,
-    lowerGraphCenterFontSize: 20,
-    lowerGraphTooltipMinWidth: 140,
-    lowerGraphTooltipFontSize: 12,
-    lowerGraphInfoBoxWidth: 240,
-    lowerGraphInfoBoxHeight: 200,
-    lowerGraphArcLabelFontSize: 12,
-    lowerGraphArcPadding: 5,
-    lowerGraphChartHeadingFontSize: 20,    
-    lowerGraphInfoBoxFontSize: 14, 
-    lowerGraphInfoBoxBorderWidth: 10,
-    lowerGraphInfoBoxPaddingLeft: 10,
-    lowerGraphInfoBoxLineSpacing: 10,
-    legendFontSize: 12,
-    legendSquareSize: 12,
-    legendRowHeight: 20,
-    tableSectionBorderLeft: "1px solid black",
-});
