--- assets/strings/strings.js
@@ -1,120 +1,0 @@
-// translations for certain text used in the project
-export const TranslationObj = {
-    en: {
-        translation: {
-            "upperGraph": {
-                "number": {
-                    "graphTitle": "Daily {{ nutrient }} intake per person ({{ amountUnit }}/d) provided by 12 food groups.",
-                    "yAxisTitle": "{{nutrient}} Intake ({{ amountUnit }}/d)",
-                    "switchTypeButton": "Switch to Percentage"
-                },
-                "percentage": {
-                    "graphTitle": "Percentage of total {{ nutrient }} intake provided by 12 food groups.",
-                    "yAxisTitle": "% of total {{nutrient}} intake",
-                    "switchTypeButton": "Switch to Numbers"
-                },
-                "graphFootnote": "Data Source: Statistics Canada, 2015 Canadian Community Health Survey - Nutrition, 2015, Share File.",
-                "tableTitle": "Absolute ({{ amountUnit }}/day) and relative (%) contribution of 12 food groups to daily {{nutrient}} intake",   
-                "infoBox_number": [
-                    "{{- name }}",
-                    "Amount: {{amount}}"
-                ],
-                "infoBox_percentage": [
-                    "{{- name }}",
-                    "{{ percentage }}% of total {{- nutrient }} intake."
-                ]
-            },
-            "lowerGraph": {
-                "graphTitle": "Food groups and sub-groups contribution to {{ nutrient }} intake in Canadian {{ ageSexGroup }}",
-                "graphFootnote": "Data Source: Statistics Canada, 2015 Canadian Community Health Survey - Nutrition, 2015, Share File.",
-                "tableTitle": "Absolute ({{ amountUnit }}/day) and relative (%) contribution of food groups and sub-groups to daily {{nutrient}} intake in Canadian {{ ageSexGroup }}, 2015",
-                "seeLevel2Groups": "Filter on level 2 groups",
-                "seeAllGroups": "See all food groups",
-                "infoBoxLevel_1": [
-                    "{{- name }}",
-                    "{{ percentage }}% of total {{ nutrient }} intake."
-                ],
-                "infoBoxLevel_2": [
-                    "{{- name }}",
-                    "Contribution to:",
-                    "Total {{ nutrient }} intake: {{ percentage }}%",
-                    "{{- parentGroup }} group: {{ parentPercentage }}%"
-                ],
-                "infoBoxLevel_3": [
-                    "{{- name }}",
-                    "Contribution to:",
-                    "Total {{ nutrient }} intake: {{ percentage }}%",
-                    "{{- parentGroup }}: {{ parentPercentage }}%"
-                ],
-                "infoBoxLevel_4": [
-                    "{{- name }}",
-                    "Contribution to:",
-                    "Total {{ nutrient }} intake: {{ percentage }}%",
-                    "{{- parentGroup }}: {{ parentPercentage }}%"
-                ],
-                /* If the context number is not between 1-4 */
-                "hoverBoxLevel_other": [ 
-                    "{{- name }}",
-                    "{{ percentage }}% of total {{ nutrient }} intake."
-                ],
-            }
-        }
-    },
-    fr: { 
-        translation: {} 
-    }
-}
-
-// text wrap attributes
-export const TextWrap  = {
-    NoWrap: "No Wrap",
-    Wrap: "Wrap"
-};
-
-// text anchor attributes
-export const TextAnchor = {
-    Start: "start",
-    Middle: "middle",
-    End: "end"
-};
-
-// font weight attributes
-export const FontWeight = {
-    Normal: "normal",
-    Bold: "bold",
-    Lighter: "lighter",
-    Bolder: "bolder"
-}
-
-// default attributes used for different components
-export const DefaultAttributes = {
-    textAnchor: TextAnchor.Start,
-    textWrap: TextWrap.Wrap,
-    fontWeight: FontWeight.Normal,
-    opacity: 1
-}
-
-// column names in the food description group csv file
-export const FoodGroupDescDataColNames = {
-    foodGroupLv1: "Food Group Name_Level 1",
-    foodGroupLv2: "Food Group Name_Level 2",
-    foodGroupLv3: "Food Group Name_Level 3",
-    description: "Description of the Contents of the Food Groups and Sub-groups"
-};
-
-// column names in the food ingredient csv file
-export const NutrientDataColNames = {
-    ageSexGroup: "Age-sex group (*: excludes pregnant or breastfeeding)",
-    foodGroupLv1: "Food group_level1",
-    foodGroupLv2: "Food group_level2",
-    foodGroupLv3: "Food group_level3"
-};
-
-// heading names for the age sex group
-export const AgeSexGroupHeadings = {
-    Population1Up: "Population age 1+", 
-    Children1To8: "Children 1 to 8 y", 
-    YouthAndAdolescents: "Youth & adolescents 9 to 18 y",
-    AdultMales: "Adult males 19 y +",
-    AdultFemales: "Adult females* 19 y +"
-};
