--- model/dataModels/csvDataModel.js
@@ -1,5 +1,0 @@
-export class CSVDataModel {
-    constructor(data) {
-        this.data = data;
-    }
-}
